requirements.txt
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
coverage==7.6.10
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
javalang==0.13.0
numpy==2.2.2
pandas==2.2.3
parso==0.8.4
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
soupsieve==2.6
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wheel==0.44.0
